+++ MoMoney.Core/ViewModels/TransactionsViewModel.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections.ObjectModel;
 using CommunityToolkit.Mvvm.Input;
 using CommunityToolkit.Mvvm.ComponentModel;
-using Syncfusion.Maui.ListView;
 using MoMoney.Core.Models;
 using MoMoney.Core.Helpers;
 using MoMoney.Core.Services.Interfaces;
@@ -34,8 +33,6 @@ public partial class TransactionsViewModel : ObservableObject
     [ObservableProperty] static DateTime to = new();
 
     List<Transaction> Transactions = [];
-
-    public SfListView ListView { get; set; }
 
     bool showValue = true;
 
@@ -107,14 +104,15 @@ public partial class TransactionsViewModel : ObservableObject
     async Task Read()
     {
         var transactions = await transactionService.GetTransactionsFromTo(From, To, true);
-        if (transactions.Count() != Transactions.Count)
+        if (transactions.Count != Transactions.Count)
         {
-            ListView.LoadMoreOption = LoadMoreOption.Auto;
-
-            LoadedTransactions.Clear();
+            // update transactions
             Transactions.Clear();
             Transactions = new(transactions);
+            LoadedTransactions.Clear();
+            await LoadMoreItems();
 
+            // update payees
             var payees = transactions.Select(t => t.Payee).Distinct();
             Payees.Clear();
             foreach (var payee in payees)
@@ -210,13 +208,13 @@ public partial class TransactionsViewModel : ObservableObject
         if (Category != null)
         {
             await GetSubcategories(Category);
-            UpdateFilter();
+            await UpdateFilter();
         }
         else
         {
             Subcategory = null;
             Subcategories.Clear();
-            UpdateFilter();
+            await UpdateFilter();
         }
     }
 
@@ -239,55 +237,22 @@ public partial class TransactionsViewModel : ObservableObject
     /// Calls UpdateFilter().
     /// </summary>
     [RelayCommand]
-    void AmountDragCompleted()
+    async Task AmountDragCompleted()
     {
-        UpdateFilter();
+        await UpdateFilter();
     }
 
     /// <summary>
     /// Updates Transactions Filter.
     /// </summary>
     [RelayCommand]
-    void UpdateFilter()
+    async Task UpdateFilter()
     {
-        if (ListView.DataSource != null)
-        {
-            ListView.LoadMoreOption = LoadMoreOption.Auto;
-            ListView.DataSource.Filter = FilterTransactions;
-            ListView.DataSource.RefreshFilter();
-        }
-    }
-
-    /// <summary>
-    /// Checks if transaction matches filters.
-    /// </summary>
-    /// <param name="obj"></param>
-    bool FilterTransactions(object obj)
-    {
-        // if all are blank, show Transaction
-        if (Account == null && Category == null && Subcategory == null &&
-            AmountRangeStart == 0 && AmountRangeEnd == 500 && Payee == "")
-            return true;
-
-        var trans = obj as Transaction;
-        var amount = Math.Abs(trans.Amount);
-        var payee = Payee.ToLower();
-
-        // if fields aren't blank and match values, show Transaction
-        if (Account != null && trans.AccountID != Account.AccountID)
-            return false;
-        if (amount < AmountRangeStart && AmountRangeStart != 0)
-            return false;
-        if (amount > AmountRangeEnd && AmountRangeEnd != 500)
-            return false;
-        if (Category != null && trans.CategoryID != Category.CategoryID)
-            return false;
-        if (Subcategory != null && trans.SubcategoryID != Subcategory.CategoryID)
-            return false;
-        if (payee.Length > 0 && !trans.Payee.ToLower().Contains(payee))
-            return false;
-        
-        return true;
+        Transactions = await transactionService.GetFilteredTransactions(
+            from, to, Account?.AccountID, AmountRangeStart, AmountRangeEnd,
+            Category?.CategoryID, Subcategory?.CategoryID, Payee);
+        LoadedTransactions.Clear();
+        await LoadMoreItems();
     }
 
     /// <summary>
@@ -296,26 +261,10 @@ public partial class TransactionsViewModel : ObservableObject
     [RelayCommand]
     async Task LoadMoreItems()
     {
-        ListView.IsLazyLoading = true;
-        await Task.Delay(250);
+        await Task.Delay(50);
         int index = LoadedTransactions.Count;
         int totalItems = Transactions.Count;
         int count = index + Constants.LOAD_COUNT >= totalItems ? totalItems - index : Constants.LOAD_COUNT;
-        AddTransactions(index, count);
-        ListView.IsLazyLoading = false;
-
-        // disables loading indicator
-        if (count == 0)
-            ListView.LoadMoreOption = LoadMoreOption.None;
-    }
-
-    /// <summary>
-    /// Copies transactions from Transactions to LoadedTransactions.
-    /// </summary>
-    /// <param name="index"></param>
-    /// <param name="count"></param>
-    void AddTransactions(int index, int count)
-    {
         var transactions = Transactions.Skip(index).Take(count);
         foreach (var transaction in transactions)
             LoadedTransactions.Add(transaction);